pyproject.toml
[tool.poetry]
name = "datas_service"
version = "0.1.0"
description = ""
authors = ["o.zamesov@mail.ru <o.zamesov@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.23"
asyncpg = "^0.29.0"
fastapi = {extras = ["all"], version = "^0.104.1"}
Jinja2 = "^3.1.2"
pytest = "^7.4.3"
pytest-dotenv = "^0.5.2"
pydantic-settings = "^2.1.0"
pytest-asyncio = "^0.23.2"
httpx = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
